requirements.txt
pygame==2.5.2
numpy==1.24.3
pandas==2.0.3
folium==0.14.0
geopy==2.4.1
python-dotenv==1.0.0
colorama==0.4.6
googlemaps==4.10.0
pillow==10.0.0
requests==2.31.0
